Counting any tests loaded by the method loadTestsFromNames() now works.
This includes django-nose and plain old nosetests. It does not count anyone who passes the `suite` kwarg directly to nose.core.TestProgram. I am going to leave that alone until somebody complains.

This commit has a lot of design notes on and wreckage of the generator-tee-ing strategy I was pursuing previously. It also sports a generalized capture_suite() which was intended for patching multiple loader methods. In a moment, I'll simplify it, but I want to preserve this general version in case I need it later.