You're so good at recognizing test frames? Fine: don't show anything before them. Closes #18.
Made frame guesser work in the opposite direction so the first thing that looks like the test wins (and we don't chop too much off the top of the traceback) if somebody has the test call into another function in the same module with the same name. Also, don't do that.

Works fine with test generators.