Never mind about the cursor hiding. It makes pdb sad.
It'd be messy to make it show when pdb takes over, because we'd have to stash the Terminal someplace that we can get to from our patched-up set_trace(). Also, there'd be no good way to show the cursor for other things, like ipdb, which is at least usable (if annoying) now.