Make "OK!" green and bold.
I've been running like this for a few weeks, and I like it. If something like Django spits out a bunch of extra crap after the main report ("Destroying databases..."), it helps me locate the test result, in the case that it went well. If it didn't go well, the bold in the error totals serves the same purpose. I'm holding off on making those red for now because red is more distracting than green and I'd have to use a lot more of it (red on the "ERROR" and "FAIL" lines to match).

Does anybody care about being able to customize this color?